# docs/openapi.yaml
openapi: 3.0.3
info:
  title: App BiT — API de Orientação Pessoal
  version: "2.0.0"
  description: |
    API REST do App BiT — orientação de carreira, saúde mental, cursos, vagas
    e infraestrutura de rede para jovens em transição de carreira tech.

    **Regra de crise (CVV):** a decisão de derivar um usuário para o CVV
    (Centro de Valorização da Vida) é **determinística no Backend**
    (`SaudeMentalService.nivelDerivacao`), baseada **exclusivamente na
    `notaSemanal` (0-10)** do check-in atual, **quando presente** — o humor e
    o texto **nunca** derivam nada, e a derivação **nunca é delegada ao agente
    de IA** (Gemini ou Mental Health Agent via n8n). A IA só gera a mensagem
    de acolhimento e a `leituraEmocional` (texto); ela nunca decide, aciona ou
    cancela o CVV. Regra graduada: `notaSemanal` 0-1 → derivação
    **REFORÇADA** (modal de consentimento antes de gravar + painel de cuidado
    com CVV em destaque); 2-3 → derivação **PREVENTIVA** (grava direto +
    escuta suave apresentada como recurso); 4-10 → **sem derivação**
    (acolhimento normal); **ausente** → **sem derivação** nesse check-in. O
    nível é devolvido em `nivelDerivacao` e `derivarCvv = (nivelDerivacao !=
    null)`.

    **`notaSemanal` é opcional (lote 4.1):** o check-in diário do Bit
    (emoji + texto livre) é enviado **sem** `notaSemanal` — o agente infere o
    estado emocional do humor/texto só para gerar `leituraEmocional`, nunca
...
    semana.

    **Convenções:**
...
  contact:
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /api/mentorias:
    post:
      tags: [Mentorias]
      summary: Cria uma solicitação de mentoria
      description: >-
        Tenta persistir no banco primeiro; se indisponível, cai no
        FallbackStorage em memória (mesmo padrão de
        `AssessmentService.salvarAssessment`). Nasce sempre com status
        `AGUARDANDO_CONFIRMACAO`.
      security: []
      parameters:
        - name: usuarioId
          in: query
          required: false
          schema:
            type: integer
            format: int64
            default: 0
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SolicitacaoMentoriaRequest'
      responses:
        '200':
          description: Solicitação criada
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponseBase'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/SolicitacaoMentoriaResponse'
        '422':
          $ref: '#/components/responses/BusinessError'
  /api/mentorias/historico:
    get:
      tags: [Mentorias]
      summary: Lista o histórico de solicitações de mentoria de um usuário
      security: []
      parameters:
        - name: usuarioId
          in: query
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Histórico do usuário, mais recente primeiro
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponseBase'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/SolicitacaoMentoriaHistoricoItem'
  /api/mentorias/{id}/status:
    patch:
      tags: [Mentorias]
      summary: Avança o status de uma solicitação de mentoria
      description: >-
        Transição sequencial única: AGUARDANDO_CONFIRMACAO → CONFIRMADA →
        CONCLUIDA (uma casa por vez, sem pular, sem voltar). Status
        desconhecido ou fora de ordem retorna 422 via `BusinessException`
        (padrão do sistema; nenhum endpoint desta API responde 400 para
        regra de negócio — só bean validation usa 400).
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AtualizarStatusMentoriaRequest'
      responses:
        '200':
          description: Status atualizado
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponseBase'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/SolicitacaoMentoriaResponse'
        '422':
          $ref: '#/components/responses/BusinessError'

  # ───────────────────────────── Vagas ─────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Token emitido por /api/auth/register ou /api/auth/login. Endpoints
        marcados com `security: []` neste documento são públicos
        (permitAll no SecurityConfig); os demais exigem
        `Authorization: Bearer <token>`.
  responses:
...
    BusinessError:
      description: "Regra de negócio violada (ex.: entidade referenciada não existe)"
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiResponseBase'
              - type: object
                properties:
                  codigo:
                    $ref: '#/components/schemas/CodigoErro'
  schemas:
    ApiResponseBase:
      type: object
      description: Envelope padrão de toda resposta da API (campos nulos são omitidos).
      properties:
        success: { type: boolean }
        error:
          type: string
          nullable: true
          description: Mensagem legível do erro (presente quando success=false)
        codigo:
          type: string
          nullable: true
          description: Código estável do erro para tratamento programático (presente quando success=false)
        timestamp:
          type: string
          format: date-time
      required: [success, timestamp]
    CodigoErro:
      type: string
      description: Códigos estáveis emitidos por BusinessException nos serviços atuais.
      enum:
        - USUARIO_NAO_IDENTIFICADO
        - CURSO_NAO_ENCONTRADO
        - VAGA_NAO_ENCONTRADA
        - MENTOR_NAO_ENCONTRADO
        - SOLICITACAO_NAO_ENCONTRADA
        - STATUS_INVALIDO
        - TRANSICAO_INVALIDA
        - EMAIL_EXISTS
        - INVALID_CREDENTIALS
        - USER_NOT_FOUND
        - N8N_UNAVAILABLE
        - WEBHOOK_ERROR
        - VALIDACAO
        - ERRO_INTERNO
        - BUSINESS_ERROR

    # ───────── Orientação ─────────
...
    CursoArea:
      type: string
      description: Taxonomia de área usada tanto por cursos quanto por vagas.
      enum: [Java, Web, Dados, Infraestrutura, DevOps, IA, Cibersegurança, Mobile, Games, UI/UX, Programação]
...
    StatusMentoria:
      type: string
      description: Transição sequencial única, sem pular nem voltar.
      enum: [AGUARDANDO_CONFIRMACAO, CONFIRMADA, CONCLUIDA]
...
    SolicitacaoMentoriaRequest:
      type: object
      required: [mentorId, areaSolicitada]
      properties:
        mentorId: { type: integer, format: int64 }
        areaSolicitada: { $ref: '#/components/schemas/CursoArea' }
        mensagem: { type: string, nullable: true }
    SolicitacaoMentoriaResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        usuarioId: { type: integer, format: int64 }
        mentorId: { type: integer, format: int64 }
        mentorNome: { type: string }
        areaSolicitada: { type: string }
        mensagem: { type: string, nullable: true }
        status: { $ref: '#/components/schemas/StatusMentoria' }
        createdAt: { type: string, format: date-time }
    SolicitacaoMentoriaHistoricoItem:
      type: object
      properties:
        id: { type: integer, format: int64 }
        mentorId: { type: integer, format: int64 }
        mentorNome: { type: string }
        areaSolicitada: { type: string }
        mensagem: { type: string, nullable: true }
        status: { $ref: '#/components/schemas/StatusMentoria' }
        createdAt: { type: string, format: date-time }
    AtualizarStatusMentoriaRequest:
      type: object
      required: [status]
      properties:
        status: { $ref: '#/components/schemas/StatusMentoria' }

    # ───────── Vagas ─────────
...